I was doing some tests regarding Occurrence (I thought I had a bug in that code, but there wasn't) and found that this multiple attribute findBy didn't work as expected: Find.ByName("Z") & new Occurrence(1) & Find.ByValue("text");
it behaved like: Find.ByName("Z") & Find.ByValue("text");

while the following findBy worked OK:
Find.ByName("Z") & (new Occurrence(1) & Find.ByValue("text"));

Subtle notation difference but a completely different result. It was a bug in the And and Or methods of the Attribute class.
I came up with a solution and added many mock tests to make sure it worked, and it does.

Also added support for the && and || syntax although WatiN already/always uses ConditionalAnd and ConditionalOr evaluation. So using & or && and | or || gives the same result.