Implemented NextSibling for FF. Note that the test NextSibling fails for IE for the following reason:
The implementation of  IE.NextSibling assumes (by casting to IHTMLElement) that the next DOM node will be an htmlelement. From a brief investigation this seems to be mostly a correct assumption, however in some cases (such as in this test) the NextSibling will not be an HTMLElement but perhaps a text node. One way to fix this would be to cast as IHTMLDomNode and check the NodeType.

For the FF implementation of NextSibling if the next node is not an HTMLElement I grab the NextSibling until we encounter a node which is an HTMLElement, making the function really NextHtmlElementSibling.

Need to decide what is correct behaviour and update both implementations if required, (could be the IE implementation is the way we want to go anyway).